Move InMemoryStateInternals to runners/core-java
