+++ runners/direct-java/src/main/java/org/apache/beam/runners/direct/CopyOnAccessInMemoryStateInternals.java
@@ -25,6 +25,12 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Map;
 import javax.annotation.Nullable;
+import org.apache.beam.runners.core.InMemoryStateInternals.InMemoryBag;
+import org.apache.beam.runners.core.InMemoryStateInternals.InMemoryCombiningValue;
+import org.apache.beam.runners.core.InMemoryStateInternals.InMemoryState;
+import org.apache.beam.runners.core.InMemoryStateInternals.InMemoryStateBinder;
+import org.apache.beam.runners.core.InMemoryStateInternals.InMemoryValue;
+import org.apache.beam.runners.core.InMemoryStateInternals.InMemoryWatermarkHold;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.transforms.Combine.CombineFn;
 import org.apache.beam.sdk.transforms.Combine.KeyedCombineFn;
@@ -34,12 +40,6 @@ import org.apache.beam.sdk.transforms.windowing.OutputTimeFn;
 import org.apache.beam.sdk.util.CombineFnUtil;
 import org.apache.beam.sdk.util.state.AccumulatorCombiningState;
 import org.apache.beam.sdk.util.state.BagState;
-import org.apache.beam.sdk.util.state.InMemoryStateInternals.InMemoryBag;
-import org.apache.beam.sdk.util.state.InMemoryStateInternals.InMemoryCombiningValue;
-import org.apache.beam.sdk.util.state.InMemoryStateInternals.InMemoryState;
-import org.apache.beam.sdk.util.state.InMemoryStateInternals.InMemoryStateBinder;
-import org.apache.beam.sdk.util.state.InMemoryStateInternals.InMemoryValue;
-import org.apache.beam.sdk.util.state.InMemoryStateInternals.InMemoryWatermarkHold;
 import org.apache.beam.sdk.util.state.State;
 import org.apache.beam.sdk.util.state.StateContext;
 import org.apache.beam.sdk.util.state.StateContexts;
